feat: rename logo-request-logs files and update imports to use camelCase
