+++ packages/app/services/index.js
@@ -5,7 +5,7 @@ const SubscriptionService = require("./subscriptions");
 const UserTokenService = require("./usertoken");
 const UserService = require("./users");
 const RequestService = require("./request");
-const LogoRequestLogsService = require("./logo-request-logs");
+const LogoRequestLogsService = require("./logoRequestLogs");
 
 module.exports = {
   ContactUsService,
